мелкие правки

+++ src/main/java/com/itgirls/bank_system/controllers/TransactionRestController.java
@@ -1,7 +1,10 @@
 package com.itgirls.bank_system.controllers;
 
 import com.itgirls.bank_system.dto.TransactionDto;
+import com.itgirls.bank_system.exception.FailedConvertToDtoException;
+import com.itgirls.bank_system.exception.UserNotFoundException;
 import com.itgirls.bank_system.model.Transactions;
+import com.itgirls.bank_system.repository.TransactionRepository;
 import com.itgirls.bank_system.service.TransactionServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -9,16 +12,20 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
+import com.itgirls.bank_system.mapper.EntityToDtoMapper;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequiredArgsConstructor
@@ -27,11 +34,12 @@ import java.util.Map;
 @Tag(name = "транзакции", description = "управление транзакциями")
 public class TransactionRestController {
     private final TransactionServiceImpl transactionServiceImpl;
+    private final TransactionRepository transactionRepository;
 
     @PostMapping()
     @Operation(summary = "добавление транзакции", description = "данный метод позволяет создать новую транзакцию")
     public ResponseEntity<?> addNewTransaction(@Valid @RequestBody TransactionDto transactionDto,
-                                               BindingResult result) {
+                                               BindingResult result, Authentication authentication) {
         if (result.hasErrors()) {
             Map<String, String> errors = new HashMap<>();
             result.getAllErrors().forEach(objectError -> {
@@ -40,8 +48,12 @@ public class TransactionRestController {
             });
             return ResponseEntity.badRequest().body(errors);
         }
-
-        return ResponseEntity.ok().body(transactionServiceImpl.addNewTransaction(transactionDto));
+        try {
+            return ResponseEntity.ok().body(transactionServiceImpl.addNewTransaction(transactionDto));
+        } catch (Exception e) {
+            log.error(e.getMessage());
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
     }
 
     @PutMapping("/{id}")
@@ -84,12 +96,15 @@ public class TransactionRestController {
 
     @DeleteMapping("/{id}")
     @Operation(summary = "удаление транзакции", description = "данный метод позволяет удалить транзакцию")
-    public String deleteTransaction(@PathVariable("id") Long id) {
+    public ResponseEntity<String> deleteTransaction(@PathVariable("id") Long id) {
+        if (!transactionRepository.existsById(id)) {
+            return ResponseEntity.badRequest().body("транзакция с id " + id + " отсутствует в бд");
+        }
         try {
             transactionServiceImpl.deleteTransaction(id);
-            return "Транзакция успешно удалена";
+            return ResponseEntity.ok().body ("Транзакция с id " + id + " успешно удалена");
         } catch (Exception e) {
-            return "Удалить транзакцию невозможно";
+            return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
 
@@ -104,7 +119,7 @@ public class TransactionRestController {
             List<Transactions> transaction = transactionServiceImpl.getTransactionByTypeOrSenderOrBeneficiary(type,
                     senderAccountId, beneficiaryAccountId);
             log.info("Найдено {} транзакций перед конвертацией", transaction.size());
-            return transactionServiceImpl.convertTransactionToDto(transaction);
+            return EntityToDtoMapper.convertTransactionToDto(transaction);
 
         }
     }
